don't overwrite file if it exists (#1900)
* don't overwrite file if it exists


* added default constructor since it's used elsewhere

+++ teku/src/main/java/tech/pegasys/teku/cli/subcommand/PeerCommand.java
@@ -13,11 +13,13 @@
 
 package tech.pegasys.teku.cli.subcommand;
 
+import com.google.common.annotations.VisibleForTesting;
 import io.libp2p.core.PeerId;
 import io.libp2p.core.crypto.KEY_TYPE;
 import io.libp2p.core.crypto.KeyKt;
 import io.libp2p.core.crypto.PrivKey;
 import io.libp2p.core.crypto.PubKey;
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -66,6 +68,13 @@ public class PeerCommand {
           int number)
       throws IOException {
     try {
+      File f = new File(params.outputFile);
+      if (f.exists()) {
+        throw new InvalidConfigurationException(
+            String.format(
+                "Not overwriting existing file %s \nDelete file or use --outputFile to point to a file that does not currently exist.",
+                params.outputFile));
+      }
       FileWriter fileWriter = new FileWriter(params.outputFile, Charset.defaultCharset());
       PrintWriter printWriter = new PrintWriter(fileWriter);
       printWriter.println("Private Key(Hex)\tPublic Key(Hex)\tPeerId(Base58)");
@@ -94,5 +103,15 @@ public class PeerCommand {
         paramLabel = "<FILENAME>",
         description = "Path/filename of the output file")
     private String outputFile = "./config/peer-ids.dat";
+
+    @VisibleForTesting
+    protected PeerGenerationParams(final String outputFile) {
+      super();
+      this.outputFile = outputFile;
+    }
+
+    PeerGenerationParams() {
+      super();
+    }
   }
 }
